Implement shardgroup drop during splits for the old shardgroups
Also did a refactor to treat the shardgroupid as an int64 throughout.
Before it has been addressed as int64/uint64/uint32, now it is only
treated as an int64.

Still miss some propagation of the creation of the new shardgroups on
all the workers to support query from any node.