Run linter to use the new editorconfig rules
It is not so easy to use a linter on all the rules as the tool I used
fails to honor .gitignore files. It ends up checking many generated
files and comes up with many false errors.

I used the following commands to check rules and attempt to fix all
problems:

- `git ls-files | xargs eclint check`
- `git ls-files | xargs eclint fix`

However, eclint was not able to fix all the problems for me
automatically. I used the check command output and fixed them by hand.