Add pytest depedencies to Pipfile
In #6720 I'm adding a `pytest` based testing framework. This adds the
dependencies for those. They have already been merged into our docker
files in the the-process repo preparation for #6720. But by not having
them on our citus main branch it is impossible to make changes to the
Pipfile, because our CI Dockerfiles and master are out of date.

Since #6720 will need some more discussion and might take a few more
weeks to be merged, this takes out the Pipfile changes. By merging this
PR we can unblock new Pipfile changes.