Fix packaging test pipelines
We had #6737 fix the same issue on main branch, but we need to fix it on
release branches as well. As the patch does not really apply to earlier
release branches, we just added the fix manually.

(cherry picked from commit 4f9a34408537397c0a567cf45c9eacaf35b088f1)