Create a DDLJob without tasklist when there are no workers in the cluster
We mark schemas as distributed objects in Citus metadata only if we need
to propagate given CREATE SCHEMA command needs to be propagated to
worker nodes. For this reason, we were not doing this for schemas that
are created while pg_dist_node is empty.

For such scenarios, we defer the schema creation to the time when user
creates the first distributed table in the schema. However, this doesn't
help for schema-based sharding because we want to sync
pg_dist_tenant_schema to the worker nodes even for empty schemas too.