Add get_relation_info hook to avoid crash adjusted partitioning
If we allow to adjust partitioning, we get a crash when accessing
amcostestimate of partitioned indexes, because amcostestimate is NULL
for them. The following PG commit is the culprit
https://github.com/postgres/postgres/commit/3c569049b7b502bb4952483d19ce622ff0af5fd6
3c569049b7b502bb4952483d19ce622ff0af5fd6

Previously, partitioned indexes would just be ignored.

Now, they are added in the list. However get_relation_info expects
the tables which have partitioned indexes to have the inh flag set
properly. AdjustPartitioningForDistributedPlanning plays with that
flag, hence we don't get the desired behaviour.

The hook is simply removing all partitioned indexes from the list.